OAuth client and user management and testing.

+++ controllers/post.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 from schemas.post import PostCreate, PostUpdate
 from models.post import Post
@@ -10,10 +11,25 @@ def create(db: Session, post: PostCreate):
     return db_post
 
 def read(db: Session, post_id: int):
-    return db.query(Post).filter(Post.id == post_id).first()
+    db_post = db.query(Post).filter(Post.id == post_id).first()
+    
+    if db_post == None:
+        raise HTTPException(
+            status_code=404,
+            detail="Post not found."
+        )
+    
+    return db_post
 
 def update(db: Session, post_id: int, post: PostUpdate):
     db_post = db.query(Post).filter(Post.id == post_id).first()
+
+    if db_post == None:
+        raise HTTPException(
+            status_code=404,
+            detail="Post not found."
+        )
+    
     post_data = post.model_dump(exclude_unset=True)
     for key, value in post_data.items():
         setattr(db_post, key, value)
@@ -22,6 +38,13 @@ def update(db: Session, post_id: int, post: PostUpdate):
 
 def delete(db: Session, post_id: int):
     db_post = db.query(Post).filter(Post.id == post_id).first()
+
+    if db_post == None:
+        raise HTTPException(
+            status_code=404,
+            detail="Post not found."
+        )
+
     db.delete(db_post)
     db.commit()
     return {"message": "Post deleted successfully"}
